Make convert_slice_2d() use raw pointers
The util function convert_slice_2d() operates on a slice using the
stride information. It will become incompatible with PlaneSlice which
will not expose a multi-rows slice anymore (see
<https://github.com/xiph/rav1e/issues/631#issuecomment-456088577>).

To keep it generic enough, we don't want to use a PlaneSlice wrapper for
every call, so make the function use raw pointers (and unsafe).