Extract copy_from_raw_u8() to a trait
In order to make Plane generic over its element type, we will need
different implementations of copy_from_raw_u8() depending of the type.

Therefore, we will need a trait that we will implement for several types
(at least u8 and u16).

Move the current implementation of this method to pave the way.