Use TileRect to fix intra prediction offsets
The previous commit fixed the confusion between absolute offsets and
offsets relative to the current tile in intra prediction. To do so, it
changed the logic to pass the whole tile as a region, with an additional
block offset.

Instead, pass the region of interest (i.e. restore the previous
behavior) but with an additional TileRect to know the current tile
location.